Add navigation link to comic detail page
Description:

Added a link or button to each comic item in the HomeView component to navigate to the comic detail page.
The link uses the router-link component provided by Vue Router and passes the comic's ID as a route parameter.
Clicking on the "View Details" link will redirect the user to the ComicDetailView component.
The ComicDetailView component retrieves the comic data based on the route parameter and displays it.